Fixed annualPay type. No changes to output, but better.

+++ Exercise 2.6.cpp
@@ -8,7 +8,7 @@ int main()
 {
     double payAmount = 2200.0;
     int payPeriods = 26;
-    int annualPay= payAmount * payPeriods;
+    double annualPay= payAmount * payPeriods;
     cout << "Your Annual Pay is: " << annualPay << "\n";
 }
